requirements.txt
Django
Pillow
dj-database-url
mysqlclient
whitenoise
djangorestframework